BoxIO : Improve the retro-fitting of pass-through to old nodes
We were using `scriptExecutedSignal()` to patch up old .gfr files as we loaded them, but unfortunately the manual setup calls have also made their way into _code_, because folks were using the old serialisation format as a learning tool. Adapt the patching so that it works for the latter case too. This has the nice side effect of avoiding `scriptExecutedSignal()`, which as we know from the previous commit, is a trap just waiting to be fallen into.