LayoutsTest : Fix error caused by app MessageHandler registration
This was the error :

```
======================================================================
ERROR: testNoPersistentLayoutsInDefaultConfigs (GafferUITest.LayoutsTest.LayoutsTest)
----------------------------------------------------------------------
Traceback (most recent call last):
  File "/home/travis/build/GafferHQ/gaffer/install/gaffer-0.53.0.0-linux/python/GafferTest/TestCase.py", line 90, in __messageHandlerCleanup
    for message in mh.messages :
AttributeError: 'ProcessMessageHandler' object has no attribute 'messages'
----------------------------------------------------------------------
```

The cause was that the call to `app.run()` was installing a new default message handler that overrode the message handler installed by the TestCase base class. To avoid this we just load the startup files without actually running the app. The alternative would be to move the message handler setup from the Application class into the `bin/gaffer.py` wrapper, but that would mean we wouldn't get the message handler set up automatically when loading and running Gaffer apps manually (for instance, in Caribou).