SceneHistoryUI : Use `Alt` rather than `Ctrl` as keyboard modifier
There have been a ton of opinions on this, and this might not be where we end up eventually, but we expect it to be an improvement over the originals as far as most folks are concerned. The main thoughts were :

- `Ctrl+` shortcuts are awkward to type, so not suitable for keypresses we expect to use frequently.
- `Ctrl+E` clashed with the shortcut for dispatching the selected executable nodes.
- Shortcuts without modifiers can be a bit too easy to hit by accident, so may not be suitable for things
  that make a big change of state.  We did consider using a row of ASD or HJK for various history operations, and it seems that something along these lines may be the best alternative candidate.
  There is an opinion that popping up a NodeEditor is too big of a state
  change though.
- `Alt+` is easy to type, because usually you're resting on Alt for access to the camera movement.

I'm sure if we ask a few more folks we'll get a few more opinions, but this is what we're going to try for now.