ProcessMessageHandler : Don't error if "frame" not in Context
To my shame, in the review for #2915, I claimed that this situation could not occur. But it does occur, specifically for PythonCommands in sequence mode, where we require the command itself to iterate over the frames. It was even I who made it this way, in 57136469af9eece77827f298388bbdaa19914fba.