ValuePlug : Replace ad-hoc hash-cache clearing with proper LRUCache
The history of this is as follows :

- We originally cleared the hash cache at the end of _every_ root computation. At this point it was just a "short term fix" for the quadratic complexity described in the removed comment.
- We discovered we could improve performance for some particularly heavy procedural scenes at Image Engine by clearing less frequently, so we could reuse more hashes. At the time we claimed a 6x speedup for a memory cost of just 100Mb.
- We discovered various problems :
	- We had scenes where the memory cost was much higher than advertised.
	- The introduction of `TaskNodeProcess` meant that for Render tasks,
	  `Process::current()` was no longer a reliable way of detecting a root
	  computation. This meant that we failed to ever clear the cache on the
	  main thread.
	- For less well-understood reasons, we could sometimes fail to clear
      the cache on a secondary worker thread. We assume this to be a result
      of a TBB task-stealing pattern where one thread ends up servicing
      tasks originated on other threads rather than completing its own root
      computations.

All credit for the discovery of the last two problems belongs with Don Boogert, who expertly diagnosed the issue but was presumably so horrified by what he saw that he promptly scarpered.

The approach taken here is to recognise that our simple ad-hoc approach had mutated into a poor-man's LRU cache with a bizarre all-at-once-every-now-and-then eviction policy. We now use IECore::LRUCache instead, meaning that our memory footprint is properly bounded, and when we do evict, we still keep things that are likely to be needed again.

Testing with a small synthetic hash-cache-heavy workload via `gaffer stats -preCache` suggests that this comes at a cost of approximately 5% in additional runtime with the default cache size. But testing with some of our heaviest production scenes shows a dramatically improved memory footprint and in several cases a decent reduction in scene generation time. For one heavy instancing setup, we have seen a 5x reduction in peak memory usage in the worst instance, a 2x reduction in peak memory usage on average, and a 35% reduction in scene generation time on average. I suspect the synthetic workload can be improved through the usage of a custom allocator, but as heavy production scenes are our main focus, the big win there outweighs the small theoretical loss.

This also paves the way for a future optimisation whereby we don't clear the entire cache when a plug is dirtied, but instead just invalidate entries belonging to that plug. In a prototype this is already showing promising improvements in interactive performance.