GafferArnold : Add ArnoldFilterMap node
This can be used to assign a filtermap shader to cameras. Our rationale for treating this as a shader assignment rather than as an item in the scene globals is that it allows different maps to be associated with different cameras easily for multi-camera rendering, and also allows interoperability with other nodes such as ShaderTweaks.

Fixes #2676.