Context : Add `EditableScope::context()` method
This provides const access to the underlying context being managed. Although the lack of this has been awkward in a couple of places, I had been avoiding adding it because I thought it would be a blocker to some context optimisation work we need to do. But in reality I think that optimisation will be a big enough break that this won't matter.