ComputeNode/ValuePlug : Introduce cache policies
This provides us with a mechanism to fix longstanding weaknesses in the interaction between Gaffer's caching and TBB tasks. To recap, our assumption was that the majority of computes in Gaffer were in themselves entirely single-threaded, not spawning any TBB tasks. We instead gained the majority of our parallelism from outer loops implemented using TBB, for instance by visiting all scene locations in parallel, or computing the tiles of an image in parallel. Furthermore, each of these parallel computes was assumed to be independent, not relying on any shared intermediate upstream values. These assumptions have never been entirely true, and have become increasingly invalid as Gaffer's usage has broadened in scope. What happened when the assumptions were broken? If parallel computes required the _same_ upstream value, then that value would be computed redundantly on each thread, and the first thread to complete would put it in the cache. And if a compute spawned TBB tasks, then we had no protection against TBB's predilection for "stealing" _outer_ tasks, and a thread waiting for an upstream parallel compute to complete might find itself suddenly performing an outer task (location or tile) which then re-requested the _same_ upstream compute it was trying to do in the first place. These two problems combined could lead to huge amounts of redundant work being performed for certain node graphs.

The new cache policies allow us to block threads that are waiting for the same upstream value, instead waiting for a single thread to do the work. And in the case of waiting for an upstream compute spawning TBB tasks, the waiting threads can now perform useful work on behalf of that compute while they wait. This is all built on top of the recent TaskMutex/LRUCache improvements.

Note that the default compute policy is currently `Legacy` rather than `Standard`. This is because the `Standard` policy requires placing the compute behind a mutex, and can lead to deadlock if an upstream task-spawning compute "steals" an outer task. Once all task-spawning computes have declared themselves with an appropriate task-based policy, we can remove the `Legacy` policy and default everything to `Standard`.