FilterResults : Improve performance using new hash/compute policies
For certain FilterResults-heavy production graphs, this yields 4-5x speedups in scene generation time.

Also move heavy computation onto internal plug and protect it from rogue contexts using `ScenePlug::GlobalScope` in the compute for the public `outPlug()`. This was necessary to avoid ContextSanitiser warnings in the tests, but may be a worthwhile optimisation in its own right if users are combining FilterResults nodes and expressions.