ShaderAssignment : Evaluate shader input using GlobalScope
For a test scene that assigns a moderately complex shader to 30000 locations, this reduces total scene generation time (compared to master) by more than 50%. With GAFFERSCENE_SHADERASSIGNMENT_CONTEXTCOMPATIBILITY=1 we still see a reduction approaching 20%, because we are able to take advantage of the compute cache. Compared to Gaffer 0.53, the reductions are even greater, being more than 60% and 30% respectively.