Arnold metadata : Don't hide geometry parameters in NodeEditor
The rationale for hiding them was that generally you wouldn't type values into them in the NodeEditor, you'd connect them in the GraphEditor. But folks were looking for them in the NodeEditor, and when they weren't there, were assuming that Gaffer didn't support them at all. The NodeEditor _can_ actually be used for making connections and querying them too, so it seems reasonable to just show them.