Examples : Register Example Gaffer files
Adds the Examples API to the GafferUI module to allow Example Gaffer files
to be registered at runtime via Python.

Examples are registered with a unique key. The key is short textual title.
If it contains ‘/‘ characters, they will be interpreted as defining parent
groups of the example. Gaffer remembers the order Examples are
registered.

Eg: `Rendering/Cameras/Setting up Anamorphic rendering` would define an
example titled `Setting up …` within a `Cameras` group under `Rendering`.

Examples can also be registered with an optional Description and/or a
list of Gaffer Node classes that are considered relevant to what is
covered by that particular file. This is to allow Users to more easily
find examples for specific Nodes.

As many Modules may need to test for the existence of their example
files, we add a convenience method to the GafferUITest.TestCase class to
verify that all registered examples exist.
This will result in duplicate checks in many common cases, but seems
the cleanest compromoise.