Render node : Don't clear cache for "scene:render:sceneTranslationOnly"
This means we still have useful cache statistics to report from the stats app, which is the main client of the "scene:render:sceneTranslationOnly" mode. This also means that the Render node no longer reports performance monitor statistics in this mode, which is likely what we want since the stats app does its own reporting.