GafferScene : Add SetVisualiser Node
Fixes #3109

The SetVisualiser follows the pattern established by AttributeVisualiser,
allowing users to visualise Set membership in the Viewer.

Membership is expressed by custom-shading objects with colours based
on the sets they are members of. If an object is a member of multiple
sets, striped shading allows all applicable set colours to be seen
simultaneously.

The Node can be focused to only consider sets in the scene that match a
particular pattern. Custom colours can be specified by the user for any
set or sets matching a pattern.

The NodeEditor provides a legend that illustrates which colors have been
used for which sets, and allows the user to easily select a sets members
or create an override for its display color.

The Node is bound to python to allow users to specify custom Viewer
Debug Shading modes in a startup file. For an example of how to do this,
see `startup/gui/viewer.py`.

All other string Plugs that reference Sets support the standard Gaffer
right-click menu helpers for inserting set names from the incoming Scene.

NOTE: This node's behaviour diverges from AttributeVisualiser in
order to better reflect the Gaffer policy that any Node that accepts a
filter, should perform no action when there is no filter connected. As
such, in order to see any meaningful results, a filter needs to be
added, even if it matches all locations.

User Facing Changes:
 - Adds SetVisualiser Node