ArrayPlug : Refuse non-array inputs
The default behaviour inherited from Plug is to accept any input as long as it has a child accepted by each of the children of the ArrayPlug. For instance, a Switch set up with a ShaderPlug and a single array element would merrily accept a Color3f output from a shader, because the input plug has a float child accepted by the ShaderPlug (plus two additional children which are ignored). This was particularly confusing for users interacting with a shader switch, because they could accidentally make a connection into the parent ArrayPlug rather than the child element they were aiming for. We tighten this up by only accepting inputs from other ArrayPlugs.

Breaking Change :

- ArrayPlug : Inputs are now required to be ArrayPlugs too.