ViewportGadget binding : Fix GIL management bugs
Both the SceneView::Camera and the CameraTool classes trigger Actions from `ViewportGadget::cameraChangedSignal()`. We must therefore ensure that the GIL is released when calling any function that might emit that signal.

This fixes hangs when editing the camera in the Viewer - I noticed it when trying to set clipping planes while a large scene loaded in the background.