Examples : Open Example files as 'untitled copies'
In order to prevent people accidentally editing Example files, they will
be opened as an 'untitled copy'. Thus requring the user to save to a new
path.

This was chosen instead of simply locking the example files, as Gaffer's
UI prevents Nodes in locked files from being edited - which would make for
a somewhat restrictive playground.

API Changes :
 - Adds the `asNew` kwarg to GafferUI.FileMenu.addScript method. When true
   the script is opened as an untitled copy.