Examples : Fix image path in contactSheet.gfr so it works with unsaved scenes.
The example made use of `${script:name}` in the image paths. This broke when
we switched to opening examples as 'new' scripts. `${script:name}` evaluates
to an empty string in the UI, but in a local dispatch, the script is saved
as 'untitled'. This means the render path does not match what the UI session
is looking for and the graph errors.

An 'immediate' fix is to hard-code the path to include `contactSheetExample`
instead of `${script:name}`. An alternative, and perhaps preferable approach
for the long-term, would be to make sure `${script:name}` also resolved to
'untitled' for un-saved scripts.  This needs some more consideration though
and we didn't want to hold up the examples release.