Serialisation : Add option not to protect the parent namespace
When serialising for cut+paste or import, we don't know if the children created by the serialisation will clash with the names of existing children when the serialisation is executed. This makes it necessary to use `addChild()` and track the children with a temporary `__children` dict. For cases where this is not necessary, we can now get a more readable serialisation by specifying the "serialiser:protectParentNamespace" variable as `false`.

Breaking Change :

- Serialisation : Added private member.