LRUCache : Support recursive getters in Serial policy
Recursion is possible when two `GetterKeys` map to the same `Key`. This provides support for scenarios such as node pass-through in Gaffer, whereby the compute for an output plug directly passes through the result from an input plug. The cache lookup for the output recurses to the cache lookup for the input, but they both share the same cache entry.