CompoundEditor : Allow tab re-arrangement and drag-drop between panels
User Changes:

 - Tabs can now be dragged to change their order as well as to move them
   between different panels.

In order to take advantage of Qt's built-in (and pretty)
`QTabBar.setMovable( True )`, we implement the entirety of tab dragging
using Qt events instead of Gaffer's drag/drop mechanism. We justified this
with the following argument:

 - We can't programmatically initiate a Gaffer drag from a Qt event.

 - Interactive re-arrangement using Qt's movable option is a better user
   experience than implementing it ourselves using only Gaffer events
   unless we re-implement the entire functionality in our own QtTabBar
   derived class so it can be suitably controlled.

 - It's effectively a private implementation to CompoundEditor.

It has limitations as an implementation in that it relies on the current
Qt implementation details - we fake mouse events during the drag to trick
the QTabBar into thinking a standard re-arrange sequence has ended.

In addition, as we're unable to pass `GafferUI.Widget` pointers around
through either the Qt or Gaffer event systems, we have to stash the
currently dragged tab in a private global. Experiments in serializing tabs
in the drag data not only looses too much state, but suffers fairly
frequent crashes when you quickly create/add/remove/destroy QWidgets to/from
a TabWidget during drag-drop. We again justify this as its a private
implementation to `GafferUI.CompoundEditor` so at least we don't leak this
nastiness.