LayoutMenu : Fix stray ref to current window layout
As there was a ref to the active CompoundEditor in the Layout menu, it
meant that it wouldn't be deleted when a new layout was loaded.

Took the opportunity to migrate the logic to a 'saveAs' method that
mirrored the existing 'save'. This prevents repeat queries to
getLayout() for each possible 'Save As: x' menu item. This isn't so bad
when it was taking references, but if we were to repr in advance its
unnecessary overhead every time the menu is open. We could use weakref,
but that would require changes to layouts.add or another wrapper
function.