CompoundEditor : Allow tabs to be de detached from the main window
This restructures CompoundEditor to aid code-reuse between its functions and
the new DetachedPanel class. This class and related management functions
aren't protected as the CompoundEditor already surfaced
`editors()`/`addEditor()`. It seemed odd to expose those allowing changes
to be made to the main window, whilst hiding the concept of detached
panels.

This also adds a ScriptWindow._TitleManager helper in order to reduce
duplication of window-title maintenance logic.  We don't want to make
this a public API at present so it remains protected.

User Changes:

Tabs can now be 'detached' from the main ScriptWindow using one of
several mechanisms:

  - Dragging a tab into a part of the screen now covered by a Gaffer
    window belonging to the same script as the starting window.

  - Choosing 'Detach <Tab Name>' in the panel menu. This will detach
    current tab into a new window.

  - Choosing 'Detach Panel' in the panel menu. This will create a new
    window with all tabs from the source panel, and collapse the parent
    splitter.

 When dragging the last tab from a detached panel window it will
 automatically close. Empty splits within a window however, are left in place.