CompoundEditor : Show correct cursor decoration when dragging tabs
Qt's DragManager manages the cursor decorations during a drag/drop operation
based on whether or not the event has been accepted. This causes problems for
us as when a tab is dragged outside a Gaffer window (or over one from another
script) then the 'forbidden' cursor is shown. In reality though, if the user
releases the mouse, we detach the tab into a new Window.

There seems to be no way to change the DragManager's behaviour, so we have
to forcibly override the cursor it is setting.

As we don't get widget events when the cursor leaves its bounds, we rely on
a sneaky global event filter (only whilst dragging) to catch all mouse moves
on the screen.

It's not perfect, as there is some call order fighting going on, so sometimes
the cursor flickers or ends up with the forbidden symbol when the mouse is
stationary, but I think its as good as we'll get without serious surgery.