PlugValueWidget : Simplify signal/slot connection handling
We were using ScopedConnection member variables in an attempt to automatically disconnect our slots when the widget was destroyed. But in Python, member variables are destroyed _after_ the instance itself, meaning it was still possible for signals to be emitted to slots containing expired WeakMethods. This was leading to exceptions being thrown when destroying layouts. We no longer need to use ScopedConnections for this purpose, because Widget now derives from Trackable, meaning the connections are cleaned up for us automatically. We do continue to use ScopedConnections where we wish to replace connections on the fly (see `__setPlugInternal()`), but are no longer relying on them for their flawed disconnect-on-widget-destruction properties.

Note that the changes in this commit fix nothing - the actual fix is that Widget now derives from Trackable. This commit just demonstrates the code simplifications arising from having automatic connection lifetime management in place. The code could be even simpler if we could omit the `scoped = False` arguments, which are necessary due to the unfortunate history of our signal bindings - see a4057f4dda02078365e97e92c387e257609cf983 for more details.