Monitor : Fix thread-safety by integrating with ThreadState
The global pool of monitors was not thread-safe with respect to concurrent processes. Registering or deregistering a monitor could cause crashes due to concurrent reads from the global monitor set, or even calls to monitors that had already been deleted. We worked around this by cancelling all BackgroundTask before modifying the active monitors, but that is ugly, and didn't work for non-BackgroundTask-based code. The upshot? We had crashes in the wild caused by the TransformTool monitor scoping clashing with the Catalogue background-save thread.

We now manage the active monitors on a per-thread basis using the ThreadState class. Fixes #2806.

Breaking Changes :

- Monitor : Removed `setActive()/getActive()` methods. Use the Scope class instead.