BackgroundTask/ParallelAlgo : Fix ThreadState handling
The split between BackgroundTask and `callOnBackgroundThread()` is a bit of an odd one, motivated partly by a desire to have a companion for `callOnUIThread()` and partly because in the rest of the Gaffer API it is _functions_ which act in the current context, not classes. In any case, the idea was that BackgroundTask didn't manage contexts, and then `callOnBackgroundThread()` layered on the transfer of context from the call site to the background thread.

But if you used BackgroundTask directly, you would get a _random_ ThreadState, inherited from whichever TBB thread decided to steal the task. That is now fixed by scoping a vanilla ThreadState in BackgroundTask. And `callOnBackgroundThread()` follows up by making sure that the current monitors are
tranferred from the call site to the background thread along with the context. This transfer is the main motivation for having Monitors be RefCounted.