CompoundEditor : Allow tab drag-drop, save detached panels with layout
User Changes:

 - Tabs can now be dragged to change their order as well as to move them
   between different panels.

 - Tabs can now be 'detached' from the main script window using one of
   several mechanisms:

   - Dragging a tab into a part of any screen not covered by a Gaffer
     window belonging to the same script as the starting window.

   - Choosing `Detach <Tab Name>` in the panel menu. This will detach
     current tab into a new window.

   - Choosing `Detach Panel` in the panel menu. This will create a new
     window with all tabs from the source panel, and collapse the parent
     splitter.

 - When dragging the last tab from a detached panel window it will
   automatically close. Empty splits within a window however, are
   left in place.

 - Layouts now store the size, position and fullscreen state of Gaffer
   windows, including detached panels windows.

 - When a layout is restored, window positions and screen affinity will
   be adjusted unless the <Alt> key is held down when the menu item is
   selected.

 - Default layouts have been updated to load in a Maximized state. Users
   can save their own default layout if this behaviour is undesirable.

Side Effects :

 - Default Gaffer layouts now open Maximized, with sensible sizes set if
   the window is un-maximized. This has the knock-on effect that docs
   screenshots are now taken from a bigger window. We need to assess if
   this is ok, or we should adjust the layout/size in screengrab.py.

API Changes :

 - The CompoundEditor split containers can now be addressed in
   stylesheets via the `#gafferCompoundEditor` selector.

 - Gaffer SplitContainers now update the `gafferNumChildren` property
   that can be used in CSS attribute queries for styling purposes.

 - Adds `Widget.currentModifiers()` that returns the currently held
   keyboard modifiers (see GafferUI.ModifiableEvent.Modifiers).

 - Adds `"tab"` and `"detachedPanel"` `GafferUI.Pointers`

 - Adds `ScriptWindow.titleChangedSignal` that is called with the window
   and updated title whenever the ScriptWindow's title changes.

Breaking Changes :

 - Layouts saved using builds of Gaffer that contain this commit will
   fail to load in previous version. If backwards compatibility is
   required, then `contrib/startup/GafferUI/layoutCompatibility.py`
   should be added to the `GafferUI` directory on the installations
   `$GAFFER_STARTUP_PATHS`.

Notes:

This is a first-pass implementation that makes the sweeping statement that
all window proeprties that are stored should always be recalled. There are
variety of situations where this may - or may not - be ideal. A concession
to this is the <Alt> key behaviour on load (leave positions alone).

Always recalling is a simper initial strategy which will hopefully surface
real-world issues fairly quickly. It'd be good to get some user testing on
this to identify any opportunities to make it more intuitive/useful.

Technical Notes:

This implementation takes an unconventional approach in order to work
around serious issues when using Gaffer or QT drag mechanisms. Please
see notes in `CompoundEditor.py` for rationale.

This is a relatively large commit, but the drag mechanism was re-written
significantly to address usability concerns identified in field testing.
As such, rebasing to correct this was a non-trivial exercise.

History is available in PR #3167.

This restructures CompoundEditor to aid code-reuse between its protected
implementation classe.

This also adds a ScriptWindow._TitleBehaviour helper in order to reduce
duplication of window-title maintenance logic. We don't want to make
this a public API at present so it remains protected for use in similar
situations where we need to maintain a window title in sync with a script's
state and path.

windowState is serialised using bottom-left NDC coordinates relative to
each screens available area. This should hopefully make for a more
portable definition (bottom-left is used over top to better match other
Gaffer conventions). We don't make use of the native Qt save/restore
geometry mechanism as it produces an opaque binary state.

The main downside to this implementation is that (as far as I can tell),
a QWidget needs to be visible in order to determine which Screen it is
on (as otherwise its windowHandle is empty). As such we can't restore
bounds properly during __init__ so it must be called in an appropriate
place after it has been placed on screen.

On OSX, the 'fix' to prevent drags breaking when the last tab is removed
don't work, as such, we have to leave a 'ghost' tab. We do our best to
hide it.