Azure : Add OSX builds
This follows the Azure reccomendations for platform-specific behaviour:

https://docs.microsoft.com/en-us/azure/devops/pipelines/scripts/cross-platform-scripting?view=azure-devops&tabs=yaml

It does mean though, that the will-be-skipped job steps still show up in
the Azure UI. If this is undesireable, we can manually pass a `platform`
into the template to stop it emitting the steps, but it does mean that
you manualy have to line up the platform switch and the target agent.
The approach taken here should mean that it just works regardless of
what you set the `pool` to.

We build on linux using a Docker container generated by GafferHQ/build.
We have to use a manual matrix as single job + strategy doesn't allow for
conditional parameters. We have to specify `container` on linux, but not
for mac. Specifying an empty container breaks validation earlier on in the
job setup phase, a valid image name breaks on mac as docker isn't available.