Stylesheet : Improve SplitContainer appearance
Putting a Widget in a SplitContainer shouldn't dramatically change its appearance over just having the same widget in a Window, but the dark background caused exactly that dramatic shift. This made the buttons row in the Local Jobs Window way too dark. I also found the gutters between Editors oppressively dark, but maybe that's a matter of taste and other folks disagree. If so, we can arrange to darken them without affecting SplitContainers everywhere.

Additionally, for some unknown reason we were drawing the vertical handle but not the horizontal one, so now we don't draw either, and the handles just appear on hover. Hopefully this is in keeping with the general direction of "cleaner, fresher" that Tom has kickstarted.