Widget : Set class name as `gafferClass` property on QtWidget
`GafferUI.Widget` will now set the `gafferClass` and `gafferClasses`
properies on its `QWidget` during initialisation, setting them to the
class name and inheritance hierarchy of the Widget.

This allows stylesheets to use the className attribute selector
(`[gafferClass=...]`) to target specific gaffer widget types. Ideally
this would be a class selector, but these are already in use by Qt.
`[gafferClasses~=...]` can be used to match all ancestors of a specific
class.

Using `gafferClass` in a stylesheet should be the preferred approach for
defining per-widget behaviour over of coupling styling to the
implementation via concrete visual properties like `gafferRounded`
etc. as they leak the _current_ style into the implementation.