Capsule : Adapt for RendererAlgo changes
Passing `nullptr` for light links means that we don't support any linking on individual objects within the capsule. This is no worse than before, because the old linking code required an EvaluateLightLinks node to generate links, and since the capsule's never used one, internal links never worked.

Longer term it would be nice to support linking within capsules, but it could be extremely difficult. Lights can be added and removed after the renderer has expanded the capsule procedural, and since the procedural contents are opaque to us, we have no obvious way of going back in and updating the links.