startup/gui/project : Don't import GafferUI modules in headless mode
This startup script is shared with the `dispatch` app. The dispatch app
runs without a gui by default (unless --gui is specified). If dispatch
is run on a machine with no X display, the imports of `GafferUI` cause
a `QXcbConnection Could not connect to display` error and a subsequent
abort.

We now only initialize bookmarks if the application sourcing the startup
file has already loaded the UI modules.