Reference : track changes to metadata on plugs
The overall goal here is to be able to promote plugs from References to Boxes
and have the metadata preserved. Previously this failed because we converted
metadata on References to be non-persistent and such metadata doesn't survive
the promotion step.

To get this working we add a mechanism that tracks edits to loaded metadata.
This allows us to determine whether metadata needs to be serialised or not so
that we can skip the step that makes all loaded metadata non-persistent.