Expression Tests: Add two additional tests showing current issues
This adds two test to both ExpressionTest for Python and OSLExpressionEngineTest.

The first test is the practical one, which shows how users can trigger incorrect results by copying and pasting.  This currently fails for Python.  The second test checks that we aren't depending on the order in which plugs are mentioned in the internal expression.  This currently fails for OSL.  The simplest solution would be to just switch Python to work like OSL where do depend on this - but that would fail for any currently existing expressions, causing a truly enormous compatibility headache.

The robust solution is to make sure both OSL and Python pass both these tests.