Expressions : Change internal convention to execute in internal form
When we need to compute the value of __out.p0 or __out.p1 on an expression node,
we need to execute the expression, to produce flat vector of outputs
corresponding to p0 and p1.  Previously, the expression we were executing would
look like "someNode.somePlug = ...; otherNode.otherPlug = ..." - making that
we were reliant on some subtle conventions to decide which was p0 and which was
p1.

However, the serialization format for the expression looks like
"__out.p1 = ...; __out.p0 = ...;".  This makes it extremely easy to identify
which output plug is which, and removes any possibility of confusion.

This commit switches things so that we execute expressions in the internal
serialization format, rather than the "external" format ( which is still
used when initially calling setExpression ).