OSLExpressionEngine : Support plugs with "__" prefix
We usually rename plugs with an extra leading underscore, but OSL forbids triple underscores,
so double underscore plugs would fail.  I now leave the prefix alone if there is already an
underscore.  I assume the purpose of the underscore is to avoid conflict with internal
identifiers - if there's already an underscore, we should be safe