When building for Houdini17.5, we want to use SideFx VDB header but not their OpenEXR.
Another change for H17.5, we don't want to use SideFx OpenEXR has they don't
ship their python binding and all their headers are together.

For openVDB on the other end, we want to use their header so we need to only
add the openVDB "generic" header when not building for Houdini because we
append SideFx header path.