SConstruct : Disable GafferCortex by default
We will no longer ship it in official public builds. The IE config keeps it enabled for Image Engine internal builds, and it's also enabled for testing on Travis. I've left it off Azure for now, because we are publishing builds from Azure, and we don't want to publish GafferCortex.