Style : Use Color4f instead of Color3f for `renderLine()` and `renderText()`
The Style class has always been awkward because it contains methods like `renderLine()` and `renderText()` which are really just generic OpenGL drawing utilities, and nothing to do with customisable styling. The rationale for making these take Color4f is that generic drawing functions should provide full control to the caller.

The rationale for keeping the Color3f arguments to the style-related methods like `renderNodeFrame()` is that the colours here are hints, and the style may reasonably want to retain the use of transparency for its own purposes.

Breaking Changes :

- Style : Changed colour arguments for `renderLine()` and `renderText()` methods. They now accept `Color4f` rather than `Color3f`.