Attributes : Don't evaluate `extraAttributes` unnecessarily
The main purpose of `extraAttributes` is to be driven by an expression, because otherwise it is more convenient to use the regular attributes plug. But Python expressions have poor performance, so to have any hope of decent performance, you would need to filter the node to limit only a few locations, commonly the root location of an asset. But `Attributes::processesAttributes()`, in its eagerness to optimise computation by returning `false`, was triggering a compute of `extraAttributes`, the very thing we're trying to avoid! We new use `isSetToDefault()` to avoid triggering unwanted computes, but still optimise the common case of not having an input connection to `extraAttributes`.