CompoundEditor : Clear button highlight when menu items chosen
The pinning/layout buttons never clear their hover highlight when items
from their menus are chosen and the menu closes. This is because the
`leave` event never makes it to the button.  This seems like the only
place we can watch the visibility of the menu.

An additional fun-fact is that menu.getVisible() returns always returns
True when called from the menu.visibilityChangedSignal.

This bug is really obvious though now we have more frequently used pinning
menu items so is worth fixing independently of the visibility test
result.