Travis : Remove GCC 4.8 testing
TL;DR : The dependencies packages we maintain via GafferHQ/dependencies are built with GCC 6, our official release compiler, the one mandated by VFXPlatform. GCC 4.8 is not compatible with these packages, so testing against them is flawed.

This was brought to light by a crash in one of the new Appleseed tests, which had the following output :

```
testAdaptors (GafferAppleseedTest.InteractiveAppleseedRenderTest.InteractiveAppleseedRenderTest) ... *** Error in `gaffer': munmap_chunk(): invalid pointer: 0x00007ffdcdff52a0 ***
======= Backtrace: =========
/lib64/libc.so.6(+0x7f5d4)[0x7ff7071aa5d4]
/disk1/john/dev/build/gaffer/lib/libGafferOSL.so(_ZNSt10_HashtableIN6IECore14InternedStringES1_SaIS1_ENSt8__detail9_IdentityESt8equal_toIS1_ESt4hashIS1_ENS3_18_Mod_range_hashingENS3_20_Default_ranged_hashENS3_20_Prime_rehash_policyENS3_17_Hashtable_traitsILb1ELb1ELb1EEEE21_M_insert_unique_nodeEmmPNS3_10_Hash_nodeIS1_Lb1EEE+0xe9)[0x7ff6aed35a49]
/disk1/john/dev/build/gaffer/lib/libIECoreAppleseed.so(_ZNSt10_HashtableIN6IECore14InternedStringES1_SaIS1_ENSt8__detail9_IdentityESt8equal_toIS1_ESt4hashIS1_ENS3_18_Mod_range_hashingENS3_20_Default_ranged_hashENS3_20_Prime_rehash_policyENS3_17_Hashtable_traitsILb1ELb1ELb1EEEE9_M_insertIRKS1_NS3_10_AllocNodeISaINS3_10_Hash_nodeIS1_Lb1EEEEEEEESt4pairINS3_14_Node_iteratorIS1_Lb1ELb1EEEbEOT_RKT0_St17integral_constantIbLb1EE+0x92)[0x7ff66bd162f2]
/disk1/john/dev/build/gaffer/lib/libIECoreAppleseed.so(+0x24bd6)[0x7ff66bd15bd6]
/disk1/john/dev/build/gaffer/lib/libIECoreAppleseed.so(_ZN15IECoreAppleseed17ShaderNetworkAlgo7convertEPKN11IECoreScene13ShaderNetworkE+0x12d)[0x7ff66bd15ecd]
```

This turned out to be because `IECoreScene::ShaderNetworkAlgo::depthFirstTraverse()` (a template function) uses `std::unordered_set` internally, and was called from both IECoreAppleseed and GafferOSL. GCC broke ABI compatibility for `unordered_set` in 4.9, and by compiling IECoreAppleseed with GCC 6 and GafferOSL with 4.8 we ran foul of that. Details of the GCC changes here :

https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=751780
https://www.gnu.org/software/gcc/gcc-4.9/changes.html

Travis testing remains in place for GCC 6 with C++14 and GafferCortex, and for OSX. There seems little reason not to move those to Azure in the future though.