NameValuePlug : Fix plug flag handling
Recap :

- Every plug has its own set of flags.
- There are no constraints on a Plug's flags in relation to the flags of its parent.
- The `Dynamic` flag is used to tag plugs that are added outside of their parent's constructor,
  meaning they will need to be recreated explicitly by the serialisation.
- Flags - especially `Dynamic` - are a constant source of bugs, and we need to get rid of
  them (but I digress).

The NameValuePlug has two main forms for its constructor. The first takes a `Data *` to define the "value" plug implicitly, and an explicit argument for the flags of the `NameValuePlug`. The second takes a `Plug
Ptr` to define the "value" plug explicitly, and then borrows the flags from that to define the flags of the `NameValuePlug` implicitly. This latter form was used in the serialisation, and led to the bugs exposed
 in the new `StandardAttributesTest` methods.

The key problem was that the serialisation was lossy : we didn't serialise the parent flags, so upon loading they were replaced with the child flags. This led to the following sequence of operations :

1. Promote a NameValuePlug from a StandardAttributes node. The promotion code sets the `Dynamic` flag to indicate that the serialisation will need to include a constructor for the promoted plug. The child plugs
 do not have the `Dynamic` flag set, which is fine, because they will be created by the parent constructor. All is well.
2. Serialise. Oops. The flags from the parent plug are omitted from the serialisation.
3. Load. Everything _seems_ OK, because the plug has been recreated, but behind the scenes the `Dynamic` flag has been lost.
4. Save. The NameValuePlug constructor is omitted from the serialisation.
5. Load. The NameValuePlug has gone, never to return.

The solution is to introduce an extended constructor which passes flags explicitly, and use it to make the serialisation lossless. It's a pity to extend the already-slightly-confusing set of constructors for Nam
eValuePlug with two more, but we can remove the lossy versions in a future major version. Alternative solutions considered were :

- Use a `const_cast` and poke the Dynamic flag into the child during serialisation. This would have fixed the bug but with downsides :
	- The serialisation would still be lossy.
	- We would be modifying state on a supposedly `const` object, and emitting signals during serialisation. I didn't fancy the gamble.
- Hack the plug promotion code to apply the Dynamic flag to the children too. This would work for plugs promoted in the future, but
  was no good for promoted plugs serialised in Gaffer 0.53. It would also mean we were specifying the Dynamic plug on children that we
  knew didn't need it.
- Monkey patch `NameValuePlug.addChild()` to transfer the Dynamic flag to the child. This would have fixed old serialisations, so could
  have been used with the above. But again, it leaves the child with a confusing Dynamic flag, and doesn't address the general problem.